Validate bytes with replace and select
